Move onEventDeleteClick to the new family event list component
